refactor(navbar): extract JS functions to module
